DECADES Simulation Test Rig Code
Collection of programs to simulate the DECADES system.

tcpListenA.py -
 Simulates TCP Server (1st data tank). Receives binary
data and converts to a human readable format, which is printed to
the console. No options are available. The IP address of the server
and the port to listen to is hardcoded and easily changed from
within the code.
To run:
	> python tcpListenA.py

tcpListenB.py -
 Simulates second TCP Server (2nd data tank), and
performs exactly the same as the first TCP server. Human readable
received data is printed to the console. No options are available. The
IP address of the server and port to listen to is hardcoded and
easily changed from within the code.
To run:
	> python tcpListenB.py

udpListenC.py -
 Simulates UDP multicast listener (as used in in-flight
visualisations). Data is expected in ASCII format and therefore is
printed  direct to the console.
No options are available. The IP address and port
to listen to is hardcoded and easily changed from within the code.
To run:
	> python udpListenC.py

faamPrtaft.py -
 Simulates the UDP multicast packets transmitted from the PRTAFT cRIO
within DECADES. The Arduino unit uses this packet to extract the flight
number and the status of the 'Weight on Wheels' (WoW) flag. The IP
address and port over which the packet is transmitted is hardcoded and
easily changed from within the code. The packet transmitted is defined
in prtaftUdp.csv.
To run:
	> python faamPrtaft.py <options>
Options:
	-n : set packet to be error pron, such that the WoWflag is
	     invalid (i.e: not a 1 or 0), the packet is truncated or
             the packet identifier is corrupted. 30% of the packets
             will be transmitted in error

faamGindat.py -
 Simulates the UDP multicast packets transmitted from the GIN cRIO
within DECADES. The packet is used to simulate the range of other
sensors within DECADES and to test whether the Arduino unit can discern
packets in the multicast stream, thus continuing to extract the flight
number and WoW flag from the PRTAFT multicast packets. The IP address
and port over which the packet is transmitted is hardcoded and easily
changed from within the code. The packet transmitted is defined in
gindatUdp.csv.
To run:
	> python faamGindat.py <options>
Options:
	-n : set packet to be error pron, such that the WoWflag is
	     invalid (i.e: not a 1 or 0), the packet is truncated or
             the packet identifier is corrupted. 30% of the packets
             will be transmitted in error