Arduino DECADES v4.1 Embedded Software (BETA)
The firmware extends version 4.0 firmware by providing additional functionality to:
- oversample the analog signal
- deactivate the backup to the on board SD card.

The back up deactivation functionality is necessary since an SD card write can take up to 600ms, leaving oversampling unattainable.
Control of the oversampling factor (must be 1: no oversampling, 2, 4, or 8) is implemented using the  in the arddCon.cfg file.
The backup deactivation is controlled by the F=? entry in the arddCon.cfg file. If F=1, then O is forced to equal 1, irrespective of the arddCon.cfg file, given the constraints outlined above.

The oversampling increases the sample rate by the oversampling factor and sums all samples over the reporting period, which in this case is fixed to 1s. The sum of the samples over this period is then divided by the oversample factor to produce an increased resolution of reported data.
An additional gain facotr of 10 is applied to the raw ADC values, so as to ensure no data is lost through the oversample division process. This gain factor is unaccounted for, so as to remain using integers only and must be dealt with as a post process.